requirements.txt
aiohttp==3.12.14
arxiv==2.2.0
beautifulsoup4==4.14.2
certifi==2025.1.31
colorlog==6.9.0
dashscope==1.23.5
duckduckgo-search==7.5.2
exa-py==1.14.9
fastmcp==2.13.0
httpx==0.28.1
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.79
langchain-deepseek==0.1.3
langchain-mcp-adapters==0.1.7
langchain-ollama==0.2.3
langchain-openai==0.3.19
langchain-tavily==0.2.3
langgraph==0.3.11
langgraph-api==0.0.30
langgraph-checkpoint==2.0.20
langgraph-cli==0.1.77
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langchain-huggingface==0.3.1
langsmith==0.3.45
mcp==1.20.0
nest-asyncio==1.6.0
numexpr==2.11.0
#openai==1.82.0
openai==2.14.0
pandas==2.3.0
Pillow>=11.2.1
playwright==1.52.0
pydantic==2.12.3
python-dotenv==1.1.0
Requests==2.32.4
streamlit==1.47.0
tavily-python==0.5.1
tenacity==9.1.2
typing_extensions>=4.14.1
pypdf==6.5.0
flagembedding~=1.3.5
huggingface-hub~=0.36.0
transformers~=4.57.3
torch~=2.9.1
modelscope~=1.33.0
numpy~=2.3.5